+++ src/app/components/dynamic-form-question/dynamic-form-question.component.ts
@@ -1,0 +1,27 @@
+import { Component, Input } from '@angular/core';
+import { FormGroup } from '@angular/forms';
+
+import { QuestionBase } from '../questions/question-base';
+
+@Component({
+	selector: 'app-question',
+	templateUrl: './dynamic-form-question.component.html',
+	styleUrls: ['./dynamic-form-question.component.css'],
+})
+export class DynamicFormQuestionComponent {
+
+	@Input() question: QuestionBase<any>;
+	@Input() form: FormGroup;
+
+	get isValid() {
+		return this.form.controls[this.question.key].valid;
+	}
+
+	get isDirty() {
+		return this.form.controls[this.question.key].dirty;
+	}
+
+	get isTouched() {
+		return this.form.controls[this.question.key].touched;
+	}
+}